requirements.txt
flask
streamlit
pandas
requests
pyodbc
python-dotenv
